# Optimize WritePlainBYTE_ARRAY: