- Pre-allocate buf as correct size
- Use a simpler API for generating little endian integers

Before this patch, `WritePlainBYTE_ARRAY` took 14% of my runtime, now it takes 4%