@alex 2020-05-10
Benchmark from #257:

```
name         old time/op    new time/op    delta
WriteCSV-44    10.2ms ± 7%     9.1ms ± 2%  -11.00%  (p=0.000 n=19+18)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    10.7MB ± 0%     9.9MB ± 0%   -7.46%  (p=0.000 n=20+20)

name         old allocs/op  new allocs/op  delta
WriteCSV-44      103k ± 0%       63k ± 0%  -39.23%  (p=0.000 n=20+20)
```